moved static methods from NSApi to NSApiTools
