+++ code/OpenDiabetes-NSApi/src/de/opendiabetes/vault/nsapi/NSApi.java
@@ -1,6 +1,9 @@
 package de.opendiabetes.vault.nsapi;
 
-import com.google.gson.*;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParseException;
+import com.google.gson.JsonParser;
 import com.mashape.unirest.http.HttpResponse;
 import com.mashape.unirest.http.Unirest;
 import com.mashape.unirest.http.exceptions.UnirestException;
@@ -24,19 +27,13 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.text.SimpleDateFormat;
-import java.time.DateTimeException;
-import java.time.LocalDateTime;
 import java.time.ZoneId;
-import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
-import java.time.temporal.ChronoField;
 import java.time.temporal.TemporalAccessor;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 import java.util.TimeZone;
-import java.util.function.BiConsumer;
-import java.util.function.Supplier;
 import java.util.logging.ConsoleHandler;
 import java.util.logging.Handler;
 import java.util.logging.Level;
@@ -52,6 +49,16 @@ public class NSApi {
      * Used to format {@link TemporalAccessor} objects in entries.
      */
     public final static DateTimeFormatter DATETIME_FORMATTER_ENTRY = DateTimeFormatter.ofPattern(DATETIME_PATTERN_ENTRY);
+
+    /**
+     * @return a SimpleDateFormat using {@link NSApi#DATETIME_PATTERN_ENTRY} with timezone UTC
+     */
+    public static SimpleDateFormat createSimpleDateFormatEntry() {
+        SimpleDateFormat format = new SimpleDateFormat(DATETIME_PATTERN_ENTRY);
+        format.setTimeZone(TimeZone.getTimeZone("UTC"));
+        return format;
+    }
+
     /**
      * {@link DateTimeFormatter Pattern} used to format dates in treatments.
      */
@@ -60,6 +67,15 @@ public class NSApi {
      * Used to format {@link TemporalAccessor} objects in treatments.
      */
     public final static DateTimeFormatter DATETIME_FORMATTER_TREATMENT = DateTimeFormatter.ofPattern(DATETIME_PATTERN_TREATMENT);
+
+    /**
+     * @return a SimpleDateFormat using {@link NSApi#DATETIME_PATTERN_TREATMENT} with timezone UTC
+     */
+    public static SimpleDateFormat createSimpleDateFormatTreatment() {
+        SimpleDateFormat format = new SimpleDateFormat(DATETIME_PATTERN_TREATMENT);
+        format.setTimeZone(TimeZone.getTimeZone("UTC"));
+        return format;
+    }
 
     static {
         LOGGER = Logger.getLogger(NSApi.class.getName());
@@ -343,7 +359,7 @@ public class NSApi {
      * @throws NightscoutDataException   if an exception occurs while exporting the data
      */
     public void postEntries(List<VaultEntry> entries, int batchSize) throws NightscoutIOException, NightscoutServerException, NightscoutDataException {
-        for (List<VaultEntry> vaultEntries : split(entries, batchSize)) {
+        for (List<VaultEntry> vaultEntries : NSApiTools.split(entries, batchSize)) {
             postEntries(vaultEntries);
         }
     }
@@ -377,7 +393,7 @@ public class NSApi {
      * @throws NightscoutDataException   if an exception occurs while exporting the data
      */
     public void postTreatments(List<VaultEntry> treatments, int batchSize) throws NightscoutIOException, NightscoutServerException, NightscoutDataException {
-        for (List<VaultEntry> vaultEntries : split(treatments, batchSize)) {
+        for (List<VaultEntry> vaultEntries : NSApiTools.split(treatments, batchSize)) {
             postTreatments(vaultEntries);
         }
     }
@@ -416,7 +432,7 @@ public class NSApi {
      */
     @Deprecated
     public void deleteEntry(VaultEntry entry) throws NightscoutIOException, NightscoutServerException {
-        deleteVaultEntry(entry, "entries", "dateString", craeteSimpleDateFormatEntry());
+        deleteVaultEntry(entry, "entries", "dateString", createSimpleDateFormatEntry());
     }
 
     /**
@@ -431,7 +447,7 @@ public class NSApi {
      */
     @Deprecated
     public void deleteTreatment(VaultEntry treatment) throws NightscoutIOException, NightscoutServerException {
-        deleteVaultEntry(treatment, "treatments", "created_at", craeteSimpleDateFormatTreatment());
+        deleteVaultEntry(treatment, "treatments", "created_at", createSimpleDateFormatTreatment());
     }
 
     private void deleteVaultEntry(VaultEntry entry, String path, String dateField, SimpleDateFormat formatter) throws NightscoutIOException, NightscoutServerException {
@@ -516,99 +532,5 @@ public class NSApi {
                 + "server status: " + status.getStatus() + "\n"
                 + "api enabled:   " + status.isApiEnabled() + "\n"
                 + "server time:   " + status.getServerTime();
-    }
-
-    /**
-     * Splits the list into multiple partitions. The Order is kept, so concatenating all partitions would
-     * result in the original list. Changes to the original list are not reflected in the partitions.
-     *
-     * @param list      list to split
-     * @param batchSize maximum size of each partition
-     * @param <T>       type of objects in the list
-     * @return list of partitions
-     */
-    public static <T> List<List<T>> split(List<T> list, int batchSize) {
-        return split(list, batchSize, ArrayList::new, List::add);
-    }
-
-    /**
-     * Splits the array into multiple partitions. The Order is kept, so concatenating all partitions would
-     * result in the original array.
-     *
-     * @param array     array to split
-     * @param batchSize maximum size of each partition
-     * @return list of partitions
-     */
-    public static List<JsonArray> split(JsonArray array, int batchSize) {
-        return split(array, batchSize, JsonArray::new, JsonArray::add);
-    }
-
-    /**
-     * Splits the list into multiple partitions. The Order is kept, so concatenating all partitions would
-     * result in the original list. Changes to the original list are not reflected in the partitions.
-     *
-     * @param list      list to split
-     * @param batchSize maximum size of each partition
-     * @param supplier  a supplier for new partitions such as {@link ArrayList#ArrayList()}
-     * @param consumer  consumer that gets a partition and an object and adds the object to the partition such as {@link List#add}
-     * @param <I>       iterable
-     * @param <T>       type of objects in the iterable
-     * @return list of partitions
-     */
-    public static <I extends Iterable<T>, T> List<I> split(I list, int batchSize, Supplier<I> supplier, BiConsumer<I, T> consumer) {
-        List<I> batches = new ArrayList<>();
-        int i = 0;
-        I batch = null;
-        for (T item : list) {
-            if (i == 0)
-                batch = supplier.get();
-            consumer.accept(batch, item);
-            i++;
-            if (i == batchSize) {
-                batches.add(batch);
-                i = 0;
-            }
-        }
-        // add last batch
-        if (i > 0)
-            batches.add(batch);
-        return batches;
-    }
-
-    /**
-     * Parses the given string as an iso date time. Applies any timezone information if possible,
-     * if no timezone is defined in the string it is treated as UTC time.
-     *
-     * @param value date and time string according to {@link DateTimeFormatter#ISO_DATE_TIME}
-     * @return object representing the given time
-     * @throws NightscoutIOException if the value cannot be parsed for any reason
-     */
-    public static ZonedDateTime getZonedDateTime(String value) throws NightscoutIOException {
-        try {
-            TemporalAccessor t = DateTimeFormatter.ISO_DATE_TIME.parse(value);
-            if (t.isSupported(ChronoField.OFFSET_SECONDS))
-                return ZonedDateTime.from(t);
-            else return LocalDateTime.from(t).atZone(ZoneId.of("UTC"));
-        } catch (DateTimeException e) {
-            throw new NightscoutIOException("Could not parse date: " + value, e);
-        }
-    }
-
-    /**
-     * @return a SimpleDateFormat using {@link NSApi#DATETIME_PATTERN_ENTRY} with timezone UTC
-     */
-    public static SimpleDateFormat craeteSimpleDateFormatEntry() {
-        SimpleDateFormat format = new SimpleDateFormat(DATETIME_PATTERN_ENTRY);
-        format.setTimeZone(TimeZone.getTimeZone("UTC"));
-        return format;
-    }
-
-    /**
-     * @return a SimpleDateFormat using {@link NSApi#DATETIME_PATTERN_TREATMENT} with timezone UTC
-     */
-    public static SimpleDateFormat craeteSimpleDateFormatTreatment() {
-        SimpleDateFormat format = new SimpleDateFormat(DATETIME_PATTERN_TREATMENT);
-        format.setTimeZone(TimeZone.getTimeZone("UTC"));
-        return format;
     }
 }
