+++ code/OpenDiabetes-NSApi/src/de/opendiabetes/vault/nsapi/NSApiTools.java
@@ -1,5 +1,6 @@
 package de.opendiabetes.vault.nsapi;
 
+import com.google.gson.JsonArray;
 import de.opendiabetes.vault.container.VaultEntry;
 import de.opendiabetes.vault.container.VaultEntryType;
 import de.opendiabetes.vault.nsapi.exception.NightscoutIOException;
@@ -8,9 +9,19 @@ import de.opendiabetes.vault.nsapi.importer.NightscoutImporter;
 import de.opendiabetes.vault.util.SortVaultEntryByDate;
 
 import java.io.*;
+import java.time.DateTimeException;
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.time.ZonedDateTime;
+import java.time.format.DateTimeFormatter;
+import java.time.temporal.ChronoField;
+import java.time.temporal.TemporalAccessor;
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
+import java.util.function.BiConsumer;
+import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
 /**
@@ -153,5 +164,81 @@ public class NSApiTools {
         } catch (IOException e) {
             throw new NightscoutIOException("Could not close stream", e);
         }
+    }
+
+    /**
+     * Parses the given string as an iso date time. Applies any timezone information if possible,
+     * if no timezone is defined in the string it is treated as UTC time.
+     *
+     * @param value date and time string according to {@link DateTimeFormatter#ISO_DATE_TIME}
+     * @return object representing the given time
+     * @throws NightscoutIOException if the value cannot be parsed for any reason
+     */
+    public static ZonedDateTime getZonedDateTime(String value) throws NightscoutIOException {
+        try {
+            TemporalAccessor t = DateTimeFormatter.ISO_DATE_TIME.parse(value);
+            if (t.isSupported(ChronoField.OFFSET_SECONDS))
+                return ZonedDateTime.from(t);
+            else return LocalDateTime.from(t).atZone(ZoneId.of("UTC"));
+        } catch (DateTimeException e) {
+            throw new NightscoutIOException("Could not parse date: " + value, e);
+        }
+    }
+
+    /**
+     * Splits the list into multiple partitions. The Order is kept, so concatenating all partitions would
+     * result in the original list. Changes to the original list are not reflected in the partitions.
+     *
+     * @param list      list to split
+     * @param batchSize maximum size of each partition
+     * @param <T>       type of objects in the list
+     * @return list of partitions
+     */
+    public static <T> List<List<T>> split(List<T> list, int batchSize) {
+        return split(list, batchSize, ArrayList::new, List::add);
+    }
+
+    /**
+     * Splits the array into multiple partitions. The Order is kept, so concatenating all partitions would
+     * result in the original array.
+     *
+     * @param array     array to split
+     * @param batchSize maximum size of each partition
+     * @return list of partitions
+     */
+    public static List<JsonArray> split(JsonArray array, int batchSize) {
+        return split(array, batchSize, JsonArray::new, JsonArray::add);
+    }
+
+    /**
+     * Splits the list into multiple partitions. The Order is kept, so concatenating all partitions would
+     * result in the original list. Changes to the original list are not reflected in the partitions.
+     *
+     * @param list      list to split
+     * @param batchSize maximum size of each partition
+     * @param supplier  a supplier for new partitions such as {@link ArrayList#ArrayList()}
+     * @param consumer  consumer that gets a partition and an object and adds the object to the partition such as {@link List#add}
+     * @param <I>       iterable
+     * @param <T>       type of objects in the iterable
+     * @return list of partitions
+     */
+    public static <I extends Iterable<T>, T> List<I> split(I list, int batchSize, Supplier<I> supplier, BiConsumer<I, T> consumer) {
+        List<I> batches = new ArrayList<>();
+        int i = 0;
+        I batch = null;
+        for (T item : list) {
+            if (i == 0)
+                batch = supplier.get();
+            consumer.accept(batch, item);
+            i++;
+            if (i == batchSize) {
+                batches.add(batch);
+                i = 0;
+            }
+        }
+        // add last batch
+        if (i > 0)
+            batches.add(batch);
+        return batches;
     }
 }
